Move initial configuration handling into Except monad
When loading the basic initialization configuration, it's useful to be able to deal with failure monadically, rather than having to chain case statements. Move the Config library into the main application (which is where it really
belongs), and used the ExceptT monad transformer over the IO monad to capture the initial setup phase. This ultimately returns a Stack () which we can run (the core load balancing application).